# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.httpError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  handler.resetPasswordRequest:
    properties:
      password:
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  handler.resetPasswordResp:
    properties:
      message:
        type: string
    type: object
...
  handler.sendResetPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handler.sendResetPasswordResp:
    properties:
      message:
        type: string
    type: object
  handler.updateImageInfoReq:
    properties:
      description:
        type: string
      gender:
        type: string
      name:
        type: string
      whale_type:
        type: string
    type: object
  handler.uploadImgResp:
    properties:
      url:
        type: string
    type: object
host: localhost:80
info:
  contact:
    name: humpback-whale
    url: https://github.com/OkDenAl/humback-whale
  description: Humpback whale catalog - it is the service for recognize humpback whale
    and store it in catalog.
  title: Humpback whale catalog service
  version: "1.0"
paths:
...
  /private/whale/update:
    post:
      parameters:
      - description: Req body json
        in: body
        name: limit
        required: true
        schema:
          $ref: '#/definitions/handler.updateImageInfoReq'
      - description: authorization bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Обновить информацию по наблюдению
      tags:
      - Наблюдение кита
  /private/whale/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: image
        required: true
        type: file
      - description: Latitude
        in: formData
        name: latitude
        type: number
      - description: Longitude
        in: formData
        name: longitude
        type: number
      - description: Saw at
        in: formData
        name: saw_at
        type: number
      - description: authorization bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.uploadImgResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Загрузить наблюдение
      tags:
      - Наблюдение кита
...
  /public/auth/reset-password:
    post:
      consumes:
      - application/json
      description: reset password using token
      parameters:
      - description: JSON
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/handler.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.resetPasswordResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Сброс пароля
      tags:
      - Аутентификация
  /public/auth/send-reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: JSON
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/handler.sendResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.sendResetPasswordResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Отправка письма со сбросом пароля
      tags:
      - Аутентификация
...
swagger: "2.0"